Move TcpServer into its own file
